requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.1.7
django-filter==25.1
django-jazzmin==3.0.1
django-svg-image-form-field==1.0.1
djangorestframework==3.16.0
idna==3.10
Markdown==3.7
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
